+++ flux/FluxCRsFrame.py
@@ -59,8 +59,8 @@ class FluxCRsFrame():
     self.autoreload_enabled = BooleanVar()
     autoreload_label = ttk.Label(self.frame_topbar, text="Autoreload:")
     autoreload_label.pack(side=LEFT, padx=5*self.style.multiplier)
-    autoreload = ttk.Checkbutton(self.frame_topbar, command=self.autoreload, variable=self.autoreload_enabled, onvalue=True, offvalue=False, takefocus=False)
-    autoreload.pack(side=RIGHT, padx=1*self.style.multiplier)
+    self.autoreload_checkbutton = ttk.Checkbutton(self.frame_topbar, command=self.autoreload, variable=self.autoreload_enabled, onvalue=True, offvalue=False, takefocus=False)
+    self.autoreload_checkbutton.pack(side=LEFT, padx=1*self.style.multiplier)
 
   def autoreload(self):
     self.autoreload_mutex.acquire()
@@ -261,32 +261,34 @@ class FluxCRsFrame():
       self.updateTable()
       self.table.heading(column_id, command=lambda: self.sortColumn(column_id=column_id, reverse=not reverse))
 
-  def search(self, text, update=True):
+  def match_found(self, text_1, text_2):
+    if text_1[0] == '!':
+      if text_1[1:] not in text_2: return True
+      else: return False
+    else:
+      if text_1 in text_2: return True
+      else: return False
+
+  def search(self, text, update=True, values_start_index=1, values_end_index=-1):
     if text != '':
       current = self.table_data
       self.table_data = []
       
       for item in current:
-        row = " ".join(list(item.values())[1:-1]).lower()
+        row = " ".join(list(item.values())[values_start_index:values_end_index]).lower()
         matched = True
         for keyword in text.lower().split(' '):
-          if keyword[0] == '!':
-            if matched and keyword[1:] not in row: continue
-            else:
-              matched = False
-              break
-          elif ':' in keyword:
+          if ':' in keyword:
             tmp = keyword.split(':')
-            if matched and tmp[1].lower() == item[tmp[0]].lower(): continue
+            if matched and self.match_found(tmp[1], item[tmp[0]]): continue
             else:
               matched = False
               break
           else:
-            if matched and keyword in row: continue
+            if matched and self.match_found(keyword, row): continue
             else:
               matched = False
               break
-
         if matched: self.table_data.append(item)
       
       if update: self.updateTable()